Changed the naming convention for database method to be suffixed "_db()" and record methods to be suffixed _"record()" for clarity.
Removed leftover code that was likely forgotten by M.E. in "db.c" while
splitting the code into "record" and "database" specific files.

The hard-wired size of the arrays in the database struct (the "MAX_SIZE"
in "db.h") makes this code require unlimited stack with static declarations.
Changed the database allocation in the "demo.c" driver to be malloc()ed to
avoid surprises to whoever tries it.

Added the "-g" switch to the CFLAGS in the makefile to trace lines with
Valgrind.
IN 2024/08/03 "INdev"